Add files via upload
The main C file that controlled the drone's controls directly from the computer. There are a bunch of other files that make up both the SDK and running program that connects to the drone, this is the one that we edited specifically for the drone to interact with our python code.